+++ web/components/connections/index.ts
@@ -12,6 +12,7 @@ export { default as DiscordDashboard } from './DiscordDashboard';
 export { default as EmailSetupWizard } from './EmailSetupWizard';
 export { default as MailDashboard } from './MailDashboard';
 export { MailClient } from './MailClient';
+export { MailAccounts } from './MailAccounts';
 export { default as CloudDashboard } from './CloudDashboard';
 export { default as CloudSetupWizard } from './CloudSetupWizard';
 export { default as WhatsAppSetupWizard } from './WhatsAppSetupWizard';